Integrate the latest Shipwright develop (LUS Fast3dGui texture-API refactor,
#6723 icon-based rando items, #6958 corrupted-save crash fix, #5838 GBK/LACS
decouple) with the V2 save-editor rework. debugSaveEditor.cpp was the only
conflicted file (10 hunks).

Resolved by keeping V2's card-based layout and equipped-slot colored borders,
while adopting develop's API, typing, and guards:
- Migrated every texture lookup to Fast3dGui via GetRawInstance + the 4-arg
  LoadGuiTexture signature; kept V2's RG_ROCS_FEATHER(_PICKER) rando textures.
- Applied #6958 hardening: guarded every itemMapping.find()->second against
  end(), and added SLOT_NONE bounds checks in DrawInventoryTab and the
  button-item helpers.
- Took #5838's RSK_TRIFORCE_HUNT_PIECES_TOTAL>0 guard (V2 styling kept).
- Kept V2's DrawEquipmentTab, small-key count/-1/MQ work, and player-state cards.
- Fixed warnings that became errors under #6733 (treat-warnings-as-errors):
  static_cast<s8> for the widened magicCapacity/ItemID fields and changed the
  GetMaxKeysForDungeon "no keys" sentinel from 255 to -1 (its int8_t value).

Advanced submodules (libultraship/OTRExporter/ZAPDTR) to the merge's recorded
commits. Builds clean (soh.exe, Release x64); clang-format clean.

+++ soh/soh/Network/Anchor/Packets/DamagePlayer.cpp
@@ -1,6 +1,5 @@
 #include "soh/Network/Anchor/Anchor.h"
 #include <nlohmann/json.hpp>
-#include <libultraship/libultraship.h>
 #include "soh/Enhancements/game-interactor/GameInteractor.h"
 
 extern "C" {
@@ -51,7 +50,7 @@ void Anchor::HandlePacket_DamagePlayer(nlohmann::json payload) {
 
     if (damageEffect == DUMMY_PLAYER_HIT_RESPONSE_FIRE) {
         for (int i = 0; i < ARRAY_COUNT(self->bodyFlameTimers); i++) {
-            self->bodyFlameTimers[i] = Rand_S16Offset(0, 200);
+            self->bodyFlameTimers[i] = static_cast<uint8_t>(Rand_S16Offset(0, 200));
         }
         self->bodyIsBurning = true;
     } else if (damageEffect == DUMMY_PLAYER_HIT_RESPONSE_STUN) {
